Make fetchMeta behave closer to "go get"
go get will fetch the HTTPS page search for the meta tag. If the meta
tag is not on the page it will then fetch HTTP and search that.

This can be useful in cases where you can intercept the HTTP traffic
easily but not the HTTPS (like github:enterprise).