build package documentation against target's architecture
Currently gddo-server builds documentation for every package by importing it
against fixed list of platforms. The import is perfomed in a round-robin
manner - if importing with goos=linux fails, it tries to import with
the next plaftorm from the list and so on until first success.

This change ensures runtime.GOOS is always first on the list, so
gddo-server always displays documentation for the current platform,
if a package happen to care.