Fix gosrc/print.go
Support for the context package was added in baa18d8, but the
gosrc/print.go debug utility was never updated, and since it has the
ignore build tag go test etc. didn't catch this.

This will not affect godoc.org operation; it's just a debug utility for
developers; it was helpful for me in examining #576.