Add Rich data for GitLab.com as code provider
When packages were searched for and disployed, platforms like GitHub and
Bitbucket had slightly better data presentation. For example the number
of stars a project has.

Additionally, the fallback Git provider that was being used for GitLab,
leverages `git ls-remote` and `git fetch` to determine what blobs are in
a tree, and what subdirectories. This combination of requests is much
slower than the 3 API calls done now.